Add -z / --show-ip support
This new option displays both the hostname and the IP address of a
host in the path.  This is particularly important when the hostname
determined by a reverse lookup has no corresponding forward record.

This is similar to the -b (both) option in tracepath, but -b was
already taken in mtr for --bitpattern.

Using this option also implies -w / --report-wide as this option isn't
terribly useful without it in report mode.

In general we endeavor to only show the IP address once per line, but
in the split view we consistently add a separate field for IP address
to the output.

Signed-off-by: Travis Cross <tc@traviscross.com>