Partial reverse of 6bb5b6b3b.
The bug in automake does not create a vulnerable distirbution. It
allows local attackers that are logged in (or running in the
background) when a distributer types "make distcheck" to possibly gain
root access on the build system. This is NOT creating a vulnerable
distribution. The fact that automake has a vulnerability does not
warrant build failures on older systems.