warnings: fix use of uninitialized warning
clang claims: variable 'port' is used uninitialized whenever switch default
is taken.  And that is a lie, but error() is not marked as function that
never returns, because it can also be used when printing warning.  So the
compiler is guessing wrong, but the easiest fix is to initialize the
variable rather than educate compiler.