asn_{open,close}: Always initialize ipinfo hash table
Currently the only pathway to asn_open() is through display_open(), so hcreate() for ipinfo was only happening if -y or -z was specified at startup.

The ability to toggle ipinfo display in interactive mode was introduced in commit d4bb85a80ee3eb76d803e3cd04de2ccdf0379415; however, no pathway to asn_open on a toggle event was provided.  This means ipinfo results were not being stored, resulting in repeated ipinfo DNS queries being made for the same IP addresses.

To fix this, always set up the hash table if there is a possibility of toggling ipinfo display.