ui: display "no route to host" error as host entry rather than abort
When a host is reported as unreachable in response to a particular
network probe, the Windows version of the mtr UI had the unfortuante
behavior of exiting and and printing a terse message.  This is
surprising to a user, who would expect mtr probes and reporting to
continue.

With this change, probing continues and a status reading
"no route to host" is displayed in the probe report.

This change will also improve the reporting of the Unix version of
mtr, now specifying "no route to host" instead of misleadingly
indicating that no reply has been received.

This fixes issue #179.