commandline: Added --interface for using a named network interface
Using '--interface' on the commandline (or '-I') will specify
a network interface by name.  This is sometimes a more convenient
alternative to using '--address' for specifying a source address
from which to send probes.

This can be useful when you have both a wired ethernet connection
and WiFi connection, and wish to use a specific connection for the
purposes of tracing.

This feature was requested in issue #207.

This change also cleans up main() slightly by factoring out the
hostent structure generation.