Fix disappearing URLs in man pages for MacOS
groff under MacOS is missing support for .UR and .UE codes, which are
used to embed URLs in man pages.  In fact, using those codes causes
the URLs to disappear from the man page entirely.

With this change, when we generate man pages, strip of the .UR and .UE
codes when building under MacOS.